Move some Station methods into shared
