+++ Content.Server/Cargo/Components/StationCargoOrderDatabaseComponent.cs
@@ -1,8 +1,8 @@
 using System.Linq;
-using Content.Server.Station.Components;
 using Content.Shared.Cargo;
 using Content.Shared.Cargo.Components;
 using Content.Shared.Cargo.Prototypes;
+using Content.Shared.Station.Components;
 using Robust.Shared.Prototypes;
 
 namespace Content.Server.Cargo.Components;
