+++ Content.Server/Commands/ContentCompletionHelper.cs
@@ -1,4 +1,4 @@
-﻿using Content.Server.Station.Components;
+﻿using Content.Shared.Station.Components;
 using Robust.Shared.Console;
 
 namespace Content.Server.Commands;
