+++ Content.Server/Maps/GameMapPrototype.cs
@@ -1,9 +1,8 @@
-using Content.Server.Station;
 using JetBrains.Annotations;
 using Robust.Shared.Prototypes;
 using Robust.Shared.Utility;
 using System.Diagnostics;
-using System.Numerics;
+using Content.Shared.Station;
 
 namespace Content.Server.Maps;
