+++ Content.Server/Nuke/Commands/SendNukeCodesCommand.cs
@@ -1,6 +1,6 @@
 ﻿using Content.Server.Administration;
-using Content.Server.Station.Components;
 using Content.Shared.Administration;
+using Content.Shared.Station.Components;
 using Robust.Shared.Console;
 
 namespace Content.Server.Nuke.Commands;
