+++ Content.Server/Station/Commands/StationCommand.cs
@@ -2,12 +2,12 @@ using System.Diagnostics;
 using System.Linq;
 using Content.Server.Administration;
 using Content.Server.Cargo.Systems;
-using Content.Server.Station.Components;
 using Content.Server.Station.Systems;
 using Content.Shared.Administration;
+using Content.Shared.Station;
+using Content.Shared.Station.Components;
 using Robust.Shared.Toolshed;
 using Robust.Shared.Toolshed.Errors;
-using Robust.Shared.Toolshed.Syntax;
 using Robust.Shared.Utility;
 
 namespace Content.Server.Station.Commands;
@@ -54,7 +54,7 @@ public sealed class StationsCommand : ToolshedCommand
     public EntityUid? LargestGrid([PipedArgument] EntityUid input)
     {
         _station ??= GetSys<StationSystem>();
-        return _station.GetLargestGrid(Comp<StationDataComponent>(input));
+        return _station.GetLargestGrid(input);
     }
 
     [CommandImplementation("largestgrid")]
