+++ Content.Server/Station/Events/StationPostInitEvent.cs
@@ -1,4 +1,4 @@
-using Content.Server.Station.Components;
+using Content.Shared.Station.Components;
 
 namespace Content.Server.Station.Events;
