+++ Content.Server/StationEvents/Events/SpaceSpawnRule.cs
@@ -1,6 +1,4 @@
 using Content.Server.Antag;
-using Content.Server.GameTicking.Rules.Components;
-using Content.Server.Station.Components;
 using Content.Server.StationEvents.Components;
 using Content.Shared.GameTicking.Components;
 using Robust.Shared.Map;
@@ -32,10 +30,8 @@ public sealed class SpaceSpawnRule : StationEventSystem<SpaceSpawnRuleComponent>
             return;
         }
 
-        var stationData = Comp<StationDataComponent>(station.Value);
-
         // find a station grid
-        var gridUid = StationSystem.GetLargestGrid(stationData);
+        var gridUid = StationSystem.GetLargestGrid(station.Value);
         if (gridUid == null || !TryComp<MapGridComponent>(gridUid, out var grid))
         {
             Sawmill.Warning("Chosen station has no grids, cannot pick location for {ToPrettyString(uid):rule}");
