+++ Content.Shared/Station/Components/StationDataComponent.cs
@@ -1,26 +1,23 @@
-﻿using Content.Server.Shuttles.Systems;
-using Content.Server.Station.Systems;
-using Robust.Shared.Serialization.TypeSerializers.Implementations;
-using Robust.Shared.Utility;
+﻿using Robust.Shared.GameStates;
 
-namespace Content.Server.Station.Components;
+namespace Content.Shared.Station.Components;
 
 /// <summary>
 /// Stores core information about a station, namely its config and associated grids.
 /// All station entities will have this component.
 /// </summary>
-[RegisterComponent, Access(typeof(StationSystem))]
+[RegisterComponent, NetworkedComponent, AutoGenerateComponentState, Access(typeof(SharedStationSystem))]
 public sealed partial class StationDataComponent : Component
 {
     /// <summary>
     /// The game map prototype, if any, associated with this station.
     /// </summary>
-    [DataField("stationConfig")]
-    public StationConfig? StationConfig = null;
+    [DataField]
+    public StationConfig? StationConfig;
 
     /// <summary>
     /// List of all grids this station is part of.
     /// </summary>
-    [DataField("grids")]
+    [DataField, AutoNetworkedField]
     public HashSet<EntityUid> Grids = new();
 }
